+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Added
-- DPS can now optionally combine summon or associate damage under the lead character with the new `Combine Associates` checkbox.
+- DPS can now optionally add summon damage to the owner's DPS row with the new `Include Summons in DPS` checkbox.
 
 ## [1.7.0] - 2026-04-12
 ### Added